split xcm-simulator
